move the url cleaning to one file
